refactor(stage-ui): move MarkdownRenderer to subfolder and add story page
