+++ packages/stage-ui/src/components/Markdown/index.ts
@@ -1,0 +1,1 @@
+export { default as MarkdownRenderer } from './MarkdownRenderer.vue'
