+++ packages/stage-ui/src/components/index.ts
@@ -2,7 +2,7 @@ export * from './DataPane'
 export * from './Gadgets'
 export * from './Graphics'
 export * from './Layouts'
-export { default as MarkdownRenderer } from './MarkdownRenderer.vue'
+export * from './Markdown'
 export * from './Menu'
 export * from './Misc'
 export * from './Physics'
